+++ packages/astro/src/container/index.ts
@@ -13,6 +13,7 @@ import { getParts } from '../core/routing/manifest/parts.js';
 import { getPattern } from '../core/routing/manifest/pattern.js';
 import { validateSegment } from '../core/routing/manifest/segment.js';
 import type { AstroComponentFactory } from '../runtime/server/index.js';
+import { SlotString } from '../runtime/server/render/slot.js';
 import type { ComponentInstance } from '../types/astro.js';
 import type { AstroMiddlewareInstance, MiddlewareHandler, Props } from '../types/public/common.js';
 import type { AstroConfig, AstroUserConfig } from '../types/public/config.js';
@@ -474,6 +475,10 @@ export class experimental_AstroContainer {
 		component: AstroComponentFactory,
 		options: ContainerRenderOptions = {},
 	): Promise<string> {
+		if (options.slots) {
+			options.slots = markAllSlotsAsSlotString(options.slots);
+		}
+
 		const response = await this.renderToResponse(component, options);
 		return await response.text();
 	}
@@ -534,6 +539,34 @@ export class experimental_AstroContainer {
 		return renderContext.render(componentInstance, slots);
 	}
 
+	/**
+	 * It stores an Astro **page** route. The first argument, `route`, gets associated to the `component`.
+	 *
+	 * This function can be useful when you want to render a route via `AstroContainer.renderToString`, where that
+	 * route eventually renders another route via `Astro.rewrite`.
+	 *
+	 * @param {string} route - The URL that will render the component.
+	 * @param {AstroComponentFactory} component - The component factory to be used for rendering the route.
+	 * @param {Record<string, string | undefined>} params - An object containing key-value pairs of route parameters.
+	 */
+	public insertPageRoute(
+		route: string,
+		component: AstroComponentFactory,
+		params?: Record<string, string | undefined>,
+	) {
+		const url = new URL(route, 'https://example.com/');
+		const routeData: RouteData = this.#createRoute(url, params ?? {}, 'page');
+		this.#pipeline.manifest.routes.push({
+			routeData,
+			file: '',
+			links: [],
+			styles: [],
+			scripts: [],
+		});
+		const componentInstance = this.#wrapComponent(component, params);
+		this.#pipeline.insertRoute(routeData, componentInstance);
+	}
+
 	#createRoute(url: URL, params: Record<string, string | undefined>, type: RouteType): RouteData {
 		const segments = removeLeadingForwardSlash(url.pathname)
 			.split(posix.sep)
@@ -587,4 +620,12 @@ export class experimental_AstroContainer {
 
 function isNamedRenderer(renderer: any): renderer is NamedSSRLoadedRendererValue {
 	return !!renderer?.name;
+}
+
+function markAllSlotsAsSlotString(slots: Record<string, any>): Record<string, any> {
+	const markedSlots: Record<string, any> = {};
+	for (const slotName in slots) {
+		markedSlots[slotName] = new SlotString(slots[slotName], null);
+	}
+	return markedSlots;
 }