+++ packages/astro/src/core/config/schema.ts
@@ -7,7 +7,7 @@ import type {
 	RemarkPlugin as _RemarkPlugin,
 	RemarkRehype as _RemarkRehype,
 } from '@astrojs/markdown-remark';
-import { markdownConfigDefaults } from '@astrojs/markdown-remark';
+import { markdownConfigDefaults, syntaxHighlightDefaults } from '@astrojs/markdown-remark';
 import { type BuiltinTheme, bundledThemes } from 'shiki';
 import { z } from 'zod';
 import type { SvgRenderMode } from '../../assets/utils/svg.js';
@@ -101,8 +101,14 @@ export const ASTRO_CONFIG_DEFAULTS = {
 		svg: false,
 		serializeConfig: false,
 		session: false,
+		headingIdCompat: false,
+		preserveScriptOrder: false,
 	},
 } satisfies AstroUserConfig & { server: { open: boolean } };
+
+const highlighterTypesSchema = z
+	.union([z.literal('shiki'), z.literal('prism')])
+	.default(syntaxHighlightDefaults.type);
 
 export const AstroConfigSchema = z.object({
 	root: z
@@ -308,7 +314,19 @@ export const AstroConfigSchema = z.object({
 	markdown: z
 		.object({
 			syntaxHighlight: z
-				.union([z.literal('shiki'), z.literal('prism'), z.literal(false)])
+				.union([
+					z
+						.object({
+							type: highlighterTypesSchema,
+							excludeLangs: z
+								.array(z.string())
+								.optional()
+								.default(syntaxHighlightDefaults.excludeLangs),
+						})
+						.default(syntaxHighlightDefaults),
+					highlighterTypesSchema,
+					z.literal(false),
+				])
 				.default(ASTRO_CONFIG_DEFAULTS.markdown.syntaxHighlight),
 			shikiConfig: z
 				.object({
@@ -600,6 +618,14 @@ export const AstroConfigSchema = z.object({
 				.boolean()
 				.optional()
 				.default(ASTRO_CONFIG_DEFAULTS.experimental.serializeConfig),
+			headingIdCompat: z
+				.boolean()
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.experimental.headingIdCompat),
+			preserveScriptOrder: z
+				.boolean()
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.experimental.preserveScriptOrder),
 		})
 		.strict(
 			`Invalid or outdated experimental feature.\nCheck for incorrect spelling or outdated Astro version.\nSee https://docs.astro.build/en/reference/experimental-flags/ for a list of all current experiments.`,