+++ packages/astro/test/custom-500.test.js
@@ -94,6 +94,7 @@ describe('Custom 500', () => {
 			const request = new Request('http://example.com/');
 			const response = await app.render(request);
 			assert.equal(response.status, 500);
+			assert.equal(response.statusText, 'Internal Server Error');
 
 			const html = await response.text();
 			const $ = cheerio.load(html);