+++ packages/markdown/remark/src/rehype-collect-headings.ts
@@ -9,7 +9,16 @@ import type { MarkdownHeading, RehypePlugin } from './types.js';
 const rawNodeTypes = new Set(['text', 'raw', 'mdxTextExpression']);
 const codeTagNames = new Set(['code', 'pre']);
 
-export function rehypeHeadingIds(): ReturnType<RehypePlugin> {
+/**
+ * Rehype plugin that adds `id` attributes to headings based on their text content.
+ *
+ * @param options Optional configuration object for the plugin.
+ *
+ * @see https://docs.astro.build/en/guides/markdown-content/#heading-ids-and-plugins
+ */
+export function rehypeHeadingIds({
+	experimentalHeadingIdCompat,
+}: { experimentalHeadingIdCompat?: boolean } = {}): ReturnType<RehypePlugin> {
 	return function (tree, file) {
 		const headings: MarkdownHeading[] = [];
 		const frontmatter = file.data.astro?.frontmatter;
@@ -59,7 +68,9 @@ export function rehypeHeadingIds(): ReturnType<RehypePlugin> {
 			if (typeof node.properties.id !== 'string') {
 				let slug = slugger.slug(text);
 
-				if (slug.endsWith('-')) slug = slug.slice(0, -1);
+				if (!experimentalHeadingIdCompat) {
+					if (slug.endsWith('-')) slug = slug.slice(0, -1);
+				}
 
 				node.properties.id = slug;
 			}